fix(java): preserve Android minified serializers
